requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
sqlalchemy==2.0.28
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
python-dotenv==1.0.1
apscheduler==3.10.4